Module malachite_q::arithmetic

Traits for arithmetic.

Modules

abs

Absolute value of Rationals.

add

Addition of Rationals.

approximate

Implementations of Approximate and ApproximateAssign, traits for approximating a Rational by a Rational with a bounded denominator.

ceiling

Implementations of Ceiling and CeilingAssign, traits for taking the ceiling of a number.

denominators_in_closed_interval

Getting all denominators of Rationals that appear in a given closed interval.

div

Division of Rationals.

floor

Implementations of Floor and FloorAssign, traits for taking the floor of a number.

is_power_of_2

An implementation of IsPowerOf2, a trait for determining whether a number is an integer power of 2.

log_base

Implementations of traits for taking the base-$b$ logarithm of a number.

log_base_2

Implementations of traits for taking the base-2 logarithm of a number.

log_base_power_of_2

Implementations of traits for taking the base-$2^k$ logarithm of a number.

mul

Multiplication of Rationals.

neg

Negation of Rationals.

next_power_of_2

Implementations of NextPowerOf2 and NextPowerOf2Assign, traits for getting the next-highest power of 2.

pow

Implementations of Pow and PowAssign, traits for raising a number to a power.

power_of_2

Implementations of PowerOf2, a trait for computing a power of 2.

reciprocal

Implementations of Reciprocal and ReciprocalAssign, traits for computing the reciprocal of a number.

root

Implementations of CheckedRoot, a trait for computing the root of a number, if the number is a perfect power.

round_to_multiple

Implementations of RoundToMultiple and RoundToMultipleAssign, traits for rounding a number to a multiple of another number.

round_to_multiple_of_power_of_2

Implementations of RoundToMultipleOfPowerOf2 and RoundToMultipleOfPowerOf2Assign, traits for rounding a number to a multiple of a power of 2.

shl

Left-shifting a Rational (multiplying it by a power of 2).

shr

Right-shifting a Rational (dividing it by a power of 2).

sign

An implementation of Sign, a trait for determining the sign of a number.

simplest_rational_in_interval

Functions for finding the simplest (lowest-denominator) Rational in an interval.

sqrt

Implementations of CheckedSqrt, a trait for computing the root of a number, if the number is a perfect square.

square

Implementations of Square and SquareAssign, traits for squaring a number.

sub

Subtraction of Rationals.

traits

Various traits for performing arithmetic operations on numbers.