Module malachite_q::conversion

Traits for converting to and from Rationals, converting to and from strings, and extracting digits and continued fractions.

Modules

continued_fraction

Implementations of traits for converting Rationals to and from continued fractions.

digits

Implementations of traits for working with the digits of Rationals.

from_bool

An implementation of the From trait for converting a bool to a Rational.

from_float_simplest

Functions for converting a primitive float to a Rational, choosing the Rational with minimal denominator that rounds to the given float.

from_integer

Implementations of the From trait for converting an Integer to a Rational.

from_natural

Implementations of the From trait for converting a Integer to a Rational.

from_numerator_and_denominator

Functions for constructing a Rational from a numerator and denominator, or from a sign, numerator, and denominator.

from_primitive_float

Implementations of traits for converting a primitive float to a Rational.

from_primitive_int

Implementations of the From trait for converting a primitive integer to a Rational.

integer_from_rational

Implementations of traits for converting a Rational to an Integer.

is_integer

An implementation of IsInteger, a trait for determining whether a number is an integer.

mantissa_and_exponent

An implementation of SciMantissaAndExponent, a trait for converting numbers to and from a mantissa-and-exponent representation.

mutate_numerator_and_denominator

Functions for mutating a Rational’s numerator and/or denominator in place.

natural_from_rational

Implementations of traits for converting a Rational to a Natural.

primitive_float_from_rational

Functions and implementations of traits for converting a Rational to a primitive float.

primitive_int_from_rational

Implementations of traits for converting a Rational to a primitive integer.

string

Implementations of traits for converting Rationals to and from Strings.

to_numerator_and_denominator

Functions for extracting or referencing the numerator and/or denominator of a Rational.

traits

Various traits for performing arithmetic operations on numbers.