Module malachite_q::exhaustive

Iterators that generate Rationals without repetition.

Structs

ExhaustiveNegativeRationals

Generates all negative Rationals.

ExhaustiveNonzeroRationals

Generates all nonzero Rationals.

ExhaustivePositiveRationals

Generates all positive Rationals.

RationalsWithDenominator

Generates all Rationals with a specific denominator and with numerators from a given iterator. Numerators that are not coprime with the denominator are skipped.

Functions

exhaustive_negative_rationals

Generates all negative Rationals.

exhaustive_non_negative_rationals

Generates all non-positive Rationals.

exhaustive_nonzero_rationals

Generates all nonzero Rationals.

exhaustive_positive_rationals

Generates all positive Rationals.

exhaustive_rationals

Generates all Rationals.

exhaustive_rationals_with_denominator_inclusive_range

Generates all Rationals in the closed range $[a, b]$ and with a specific denominator, in order of increasing absolute value.

exhaustive_rationals_with_denominator_range

Generates all Rationals in the half-open range $[a, b)$ and with a specific denominator, in order of increasing absolute value.

exhaustive_rationals_with_denominator_range_to_infinity

Generates all Rationals greater than or equal to some number $a$ and with a specific denominator, in order of increasing absolute value.

exhaustive_rationals_with_denominator_range_to_negative_infinity

Generates all Rationals less than or equal to some number $a$ and with a specific denominator, in order of increasing absolute value.