Function malachite_q::exhaustive::exhaustive_positive_rationals

pub const fn exhaustive_positive_rationals() -> ExhaustivePositiveRationals 

Generates all positive Rationals.

The Rationals are ordered as in the Calkin-Wilf sequence. Their numerators and denominators are given by the Stern-Brocot sequence. To generate the latter sequence, this iterator uses the formula $$ a_{n+1} = \left ( 2 \left \lfloor \frac{a_{n-1}}{a_n} \right \rfloor +1 \right ) a_n - a_{n-1}, $$ attributed to David S. Newman at https://oeis.org/A002487.

The output length is infinite. The numerators and denominators of the $n$th element are $O(n^\frac{\log \phi}{\log 2})$.

Worst-case complexity per iteration

$T(n) = O(\log n \log\log n \log\log\log n)$

$M(n) = O(\log n \log\log n)$

where $T$ is time, $M$ is additional memory, and $n$ is the iteration number.

Examples

use malachite_base::iterators::prefix_to_string;
use malachite_q::exhaustive::exhaustive_positive_rationals;

assert_eq!(
    prefix_to_string(exhaustive_positive_rationals(), 20),
    "[1, 1/2, 2, 1/3, 3/2, 2/3, 3, 1/4, 4/3, 3/5, 5/2, 2/5, 5/3, 3/4, 4, 1/5, 5/4, 4/7, 7/3, \
    3/8, ...]"
)